pyproject.toml
[tool.poetry]
name = "example_api"
version = "1.0.0"
description = ""
authors = ["Dustin Cook <dustin.kenneth.cook@com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.1"
gunicorn = "^20.1.0"
uvloop = "^0.16.0"
anyconfig = "^0.13.0"
toml = "^0.10.2"
importlib-metadata = "^4.11.4"

[tool.poetry.dev-dependencies]
pytest-aiohttp = "^1.0.4"

[tool.poetry.scripts]
example_api = "example_api.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
